src/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.16
azure-nspkg==3.0.2
azure-storage==0.36.0
boto3==1.9.33
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
docutils==0.14
future==0.17.0
idna==2.7
ijson==2.3
jira==2.0.0
jmespath==0.9.3
oauthlib==2.1.0
pbr==5.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodomex==3.7.0
pyflakes==2.0.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
s3transfer==0.1.13
six==1.11.0
snowflake-connector-python==1.7.0
snowflake-sqlalchemy==1.1.2
SQLAlchemy==1.2.12
urllib3==1.24
